refactor: move the RatingSystemAdapter into the user module
This adapter is only needed in the user module, so move it into there while its only used there
